producer/requirements.txt
kafka-python==2.0.4
requests==2.31.0
schedule==1.2.1
python-dotenv==1.0.1